Vector Pencil Logo
While there are a lot of other changes in this commit, the main one is that the pixel art pencil logo for Smart Text Editor now has a vectorized counterpart! It is only used within the header, adjacent to the title.
Also, the secret Display view seems to have been removed...