Editing Tools (Not fully functional in Firefox)
This update adds a new division to the card system: Editing Tools! ...a new widget that allows touch devices to perform actions that would otherwise be difficult without having access to the keyboard shortcuts available to desktop users. The current implementation is based on the aging document.execCommand(), so Firefox users won't be able to use all of the current Editing Tools available. Hopefully a new API will be implemented for text manipulation in textareas so all browsers can be supported easily!