Scaler Visual Bug Fix
Fixed a very old issue that's been around for a long time that would cause the scaler to snap to the end position of the transition when changing to Split View or a different Orientation. This was caused because the hover state of the scaler had an incorrect property that would only allow it to transition it's background color, while it still needed to finish it's movement transition.