File Import Fix
Attempting to import a file two or more times in a row wouldn't behave as expected, the file would only import on the first try because the file input event handler would only fire when the file input received a new value that wa different than what it contained before a file was attempted on being imported. This was fixed by clearing the file input's value after a file was successfully or attempted to be imported.