Color Scheme Alignment
This commit adjusts various interface colors to look better together on screens with standard brightness. The original color scheme was developed on a non-color accurate screen, so I wanted to change the appearance to look better on screens with correct color accuracy. The design used to be very dull! This update adds a little light to the dark mode theme!