Close Button Improvements + Organization
This update simplifies buttons that are based around SVG images to be sub versions of the standard button appearance, as they are fairly similar anyways. Also, close buttons now appear on the left side of their containers when using the text editor on an Apple device! Many other changes to simplify preexisting implementations were made as well.