Better Anchor Tags and Modal UX
This commit changes all anchor tags to have no underline until they are hovered or focused upon. The other change is the removal of the automatic closing of modal cards when changing view type because this seems to be more confusing than how it was implemented previously (not closing the modal).