Modal Focus Looping
This commit updates active modals to contain the current element focus from escaping to other visible content! To escape this loop, the modal must be closed. Before this update, elements not accessible with the cursor could be navigated to using Tab and Shift Tab keyboard navigation, an abnormal result compared to what the user would normally expect to occur.