First Loading Overlay Appearance + Transition Performance Enhancements
This update adds the showing of the loading overlay when resetting all settings from the settings modal! Other changes were made as well, mostly having to do with how elements are sized when they are not active. This improves performance drops on some devices that struggle with expensive transitions and animations.