The Big Refactor: Initial Reworks
Hey, it's Brandon from the future! The next 25 commit messages having to do with "The Big Refactor" before the full release were rewritten a while after the commits themselves were authored to the repository (May / November of 2021, when this was written as well), so remember that the point of view of each commit description will be from after the refactor has already been released to the public.

This is the first commit, of many, that will drastically change the source of the site as part of "The Big Refactor". The previous commit prior to this one was the last main update to the site that was not part of the refactor, and it was the source for the public version of the site up until the refactor was released.
If you would like to see the "relatively complete", or first official release of the refactor, jump ahead to the commit named "The Big Refactor: Full Release!" to see how much the refactor accomplished. Most of the commits between here and there are not complete versions of the site, as each feature was rewritten one at a time with each commit, so be prepared for jankiness if you stop at one of those along the way! I definitely recommend checking some of them out, each little step is an important piece to making the entire puzzle of the refactor work!
To load Numbered-Textarea-Plugin along the way to get all of the site's functionality, you can clone Num-Text-Component, checkout the repo on the commit named "Compressed Various Functions" from December 13, 2020, start a Live Server hosting that folder on Localhost, and then as you go back and forth between commits in STE, modify the Numbered-Textarea-Plugin sources to use the Localhost source instead! Sorry for the little hassle, I think instances like these are where npm would be useful. I plan to look into making Num-Text-Component installable with npm, then this shouldn't as much of an issue.

Breakdown for this commit:

Additions:
- Added .config as a preapproved file extension.

Changes:
- Renamed many element's IDs with camel-case formatting to use underscores instead. I thought this would help with readability when determining whether a function is referencing a variable compared to referencing an element's ID. Example of what changed: <input id="userInput"> to <input id="user_input">

Removals
- Removed the existing custom Theme settings support, as it will be re-implemented at a later date with more user-accessible customizability using CSS instead of individual form controls.
- Removed pointer cursors from various elements, as native desktop applications tend to use the default cursor on various elements instead of the pointer cursor, such as buttons and other form controls.
- Removed the Screen Density Scaling setting.