The Big Refactor: Full Release!
This is the first public release of the new site refactor! The site has a completely new design scheme, and all JavaScript and CSS have been rewritten from the ground up. Many features have not been reimplemented yet, but I think it will make it easier to find issues when I can deploy them to all of my devices at once, rather than stick around with the previous classic version. A bug-less future is in development!