Re-redesign pt. 1
The last update was a little too drastic of a design change for me after seeing it in action, so I will be working to combining elements from the new design with the classic design, as I think the last design really had something special about it that shouldn't be left behind.
This is part 1 of many updates to come before the update is rolled out to the official site, I was a little ahead of myself when I published it out to use officially, there are still many issues that need to be fixed first.