Updated Icon Set pt. 2
Modernized both the Apple Touch Icon, and the Maskable PWA app icon. I also created a large Apple Touch Icon that may eventually be used for macOS installation if app icons can be modified based on the system that is using the site, as the current app logo without a background does not match the design aesthetics used by other apps updated for macOS Big Sur, which commonly have backgrounds.