macOS Big Sur App Icons
I figured out how to dynamically change the Web App Manifest when the page loads, so I finally have the ability to make a better icon set that matches the styles of macOS Big Sur! Figma is a really great app, it's all I used to make the new icons! Well, other than STE to stylize the output SVG code :)