Refresh Preview Changes
The refresh preview function now behaves similar to how it used to, as it is much faster than the newer way I was testing with. However, the undo stack breaks on mobile and desktop Safari when refreshing for some reason. I hope I will be able to use this method without losing undo/redo functionality, that would be the best scenario.