Platform Specific App Icons
A working version of a way to dynamically customize the Web App Manifest without a server! This is done by adding a modified version of the manifest to the page after declaring what info should be used when the app is used on their system. I added this to be able to have an icon set specific to macOS Big Sur, as the app's current icon set doesn't really match the new app icon designs that were added.

Also, I fixed any issues that may have occurred with the service worker, and the site should now meet the new requirements for installation that take effect in Chrome 93. Check the link below for more information about the new PWA installation criteria.
https://goo.gle/improved-pwa-offline-detection