SW Powered Responsive App Manifest
The App Manifest's responsiveness in regards to site origin and platform-specific icons are now managed by the service worker! This allows the modified manifest to be cacheable and to be modified before the default manifest is sent to the client.

Another change to note: The site is no longer installable before the service worker has had a chance to load. The site is only installable after the service worker has provided the complete App Manifest for the client to read.