Main Site Description
Updated the main site description! I thought the description I wrote for the ProductHunt post turned out really well, so I took part of it to use for the main site description too. The Open Graph description stayed the same as before, because I thought that it was the right length for Open Graph use cases, since most uses have small content areas and not much room for a long string of text.