Stricter Variable Interactions + Web Share Target Removal
Thanks to Vio Cassel for the Pull Request! He suggested strengthening some of the variable definitions and type checking within the Service Worker. I haven't looked too closely into those areas of JS yet, I think it would be a nice improvement to my scripts overall to start using those. Once I know how they work a little more, I'll go through the rest of the project to update any other places that could use the same help.

Here's a link to the Pull Request for the parts of this commit suggested by Vio Cassel:
https://github.com/Offroaders123/Smart-Text-Editor/pull/8

I also removed the Web Share Target information for now, I may re-add it later when I have a change to be able to test it on devices that support it.