App Omnibox Context Menu with WCO
Temporarily fixed an issue where the context menu for the Smart Text Editor on GitHub omnibox button would appear below the Window Controls Overlay inset on browsers that support it. Right now the context menu will have a minimum left and right offset value based on how large the left and right window insets are. The one issue I still have to fix is that these minimum offsets get added no matter where you are opening the context menu on the y-axis, so will not let you show the context menu all the way on the left or all the way on the right of the screen, if any window insets are present on the page for that side.