WCO Geometry Change Detection
Temprarily improved the current system I had for detecting changes in
the Window Controls Overlay state. A lot of the app feels like a lot of
spaghetti code as of now, so this will be a fix until much nicer
detection is put in place. I had it more difficult than it needed to be
previously. There is a dedicated `geometrychange` event under the
`navigator.windowControlsOverlay` property, and this is much more
reliable than simply detecting resize events on the window, which is
what I was using before, for some reason.