Testing TypeScript
About to start experimenting with a little rewrite, in TypeScript! First I'm just going to try and get the Service Worker set up, as that has been one of the harder things to get going when working on STE previously. Having types for the Service Worker will really help out so much! I'm really glad that Promises and using Async-Await makes more sense to me now also, as that was my biggest challenge for understanding the Service Worker lifecycle previously.