GPL License Upgrade
Found a video of Linus Torvalds talking about the difference between GPL version 2 and 3, and I didn't realize the stark difference between the two! GPL v2 is more what I was looking for I guess, so I'm "upgrading down" to the other version haha.

https://www.youtube.com/watch?v=PaKIZ7gJlRU

I'm not an expert on licensing, but this sounded like a nice move. I think the main difference that v2 has over v3 is that it allows the original project to use changes from a fork that diverged from the original project. It sounds like with v3, you can't do that. Interesting! I definitely want to learn more about the strength and rules that each of the common licenses out there have. MIT is really nice too, I love the open nature of the "do anything you want with it" kind of thing, haha.