App Script Typing
WOW, that was a lot of typing, as in type checking haha. Been working on this for, I'm not sure how many hours now. Listened to Yellow Hedgrerow Dreamscape two times in a row, all the way through. And I listened to Mike Keneally Dog (that was first actually), second half of Dogman (that was first first), then The Divine Wings of Tragedy 1.5 times through. Now I'm on to Demiurgs of Plancklengthville. Gonna add some legacy type definitions for my old versions of Num Text and Menu Drop, which are in use by STE right now. That should help with the final few ~120 type errors left. I'm actually surprised how much of the codebase I tackled today! I didn't redo anything, just typing out what's already there. There were so many errors before, it was really bad. Then I can start working on a nice rehash of the whole app. I'm only gonna do that with type safety though, and that's probably gonna be for another day. That would be too much to pack in after all of these changes. So happy STE is getting fixed up again!