Local Storage JSON
Found a type coalescing bug which was relying on converting `null` to a `"null"` string, and then `JSON.parse()` was parsing that string, which was valid in runtime, but not in TypeScript. To initially fix the type error, I added a nullish coalescing check which resolved to an empty string. But, since that's not JSON-parseable, it caused the `Editor.js` script to crash, which threw off the next script `app.js` when it tried to run, too.

I noticed this the other day when trying to run the app, after I added all of the type definitions, and it didn't work, to my own surprise. Guess this is one of those rare instances where TypeScript was "wrong", in a sense kind of; But only because I was doing it not the best way in the first place, then I corrected the type issue with something that was actually wrong haha.

After this fix, I noticed that the Workspace View and Orientation change transitions appear to be broken, and I haven't figured out what that's all about yet.