Spacing Formatting
My old code formatting was not the prettiest! I thought that the smaller the code was, the easier it was to read all at once. Now, I like to give my code more room to breathe, so I did that to this codebase to help it out a little bit :)