Workspace Animation Timings Fix
From refactoring a bit of animation timing logic for various workspace animations, turns out it actually change the resulting logic, causing the animations to break. This brings the code back to using the same calculations it was before the types rework. So, now the animations work again! There's are much nicer ways to animate things, now that I know better (haha), but refactoring it to use a new way is for a different time. Gotta get it working again first!

This was the only visibly breaking change from all of the typing reworks, so I wanted to figure this one out before doing my new goal, of moving all of the type safety additions to the new branch.

Being able to see the current state of your project, active in production, can help you to find things you may have missed during the new changes. I don't recommend doing it for larger projects, but I think taking those kinds of leaps for STE are what make it so strong! Gotta try out the actual code out in the real world, see if it holds up to your own standards when you're away from being able to fix things right away :)