App Script Modulation!
Moving the codebase into using ESM! Doing this in stages, so I'm uncovering some oddities for how things have been set up for a long time now.

With moving to modules, I can no longer rely on the app's functions to be accessible on the window anymore, meaning I have to manually add them to the global scope if I am to call them from the app's HTML attribute event handlers, which is how things are and have been set up since the creation of the app.

I think the fix for this is to more componently-link all of the app's functionality with it's HTML/elements, likely using Web Components in more places, now that I know how to use them to build cool things :)

Had to move the `setSyntaxHighlighting()` function from the main App script into the STE script, as I couldn't import both that function into the STE script, and also the `STE` object into the App script. It looks like that was the only module-related reference issue, at least that I've found so far. Everything appears to be working the same so far, which is great! The cross-importing reference issue is the same thing that happened to a bunch of code when starting to move Flatlands to use ESM, also.