JSON Syntax Error Message
Fixed the JSON error message parsing so it doesn't silently error in browsers with less-specific error messages. The error parsing logic was depending on all messages having the line information in the error message, and this would cause silent errors for the user when trying to format JSON with a syntax issue in it.