App Component Modules
Now that the codebase is fully typed out and documented, I can "safely" start moving parts of the main App script into separate modules, specific to each part of the app's functionality.

This is already really helping me with trying to better visualize how the code works again. Having TypeScript linting in my editor for errors is amazing too, it's helping out so much for this to see when references to things aren't there like they should be.

Separating everything into modules is also helping me come up with ideas on how I can refactor and modernize the front end for the app, Web Components, classes, and Promises with async-await being the most helpful things that will help make the app more robust.