Editor Component!
This is a big one! Been working on it all day.

This merges the Editor functionality into a single component, currently just an object/class which will instead store the data for the given Editor, rather than relying on getting that data from the corresponding elements in the DOM, which has shown to be veerry messy. This is one step I've been wanting to take with this project for a while now, I think it really helps containerize/scope the logic for the different parts of the app. Next I will make it so that this Editor class is just an extension of `NumTextElement`, meaning it's just a single Web Component which internally manages it's own state, so you don't have to worry about how it works.

This is the "last commit" on my Chromebook, at least before the new wave has come around. Ordered the base model 2021 MacBook Pro refurbished yesterday, with $8 one-day shipping, and it got here today! Was partway through a lot of this rework once UPS got here, haha. Heard the truck door from my room XD  Gonna make the next commit to my GitHub from the new Mac, it's all set up now! Transferred everything from my iMac using a Time Machine backup, and Migration Assistant. Appears to have worked amazingly so far! The only hiccup is that I'm having to manually remove and reinstall any apps/binaries/packages that are x86-64, and replace them with their Apple Silicon ARM counterparts. Alongside that, had to reinstall Homebrew and everything I had set up with that. But, all of my settings for all of the apps were still intacked, and that was my biggest thing that I was hoping I wouldn't have to deal with. Horray!

Alright, taking a rest tonight after a long tech day, gonna have a second one tomorrow, now that everything is all set up and nearly ready to go. Time to make some more cool sh!ti Thanks Theo, haha.

Oh yeah, tried vanilla Minecraft 1.19 with the native ARM build, got 102 fps while running 120hz, f yeah that's cool!