Window Controls Overlay + Editor Typing Bugfinds
Removed the old manual implementation for working with the previously experimental Window Controls Overlay PWA feature. Now that it's a stable feature supported in all modern Chromium implementations, I can rely on the final implementation of the API, which wasn't completely finished when I added support for it into STE.

I initially had to do a workaround to support `@media (display-mode: window-controls-overlay)`, which now appears to be fully supported! No need to add my own class to manage that on the document, just a few simple CSS Media Query selectors will do :)

Now that everything in the codebase is fully typed-out, I increased the type specifics for the `STE.query()` function, as it was missing a few `null` return types, causing unexpected errors for some of it's applications where it didn't expect to come across `null` values, when it in fact was working with some. This fixes the error of going back to the main Settings card, which would error out if there weren't any Editors to focus while managing the previously-focused element state. Now that the return type for `STE.query()` is correctly typed to possibly not return anything, now it will only try to focus the given Editor (if one exists).

Adjusted the `setTitle()` function to now format the app title to have the app name before the app label, which is how the PWA header is shrunk in Chromium's standalone mode. Essentially, if the window's title starts with the same as the app's name, it will only show the name once, rather than prefixing the title with the app name a second time. So, like this, now STE uses the second one:

```
Chromium will prefix with the app name and a dash, only if the title doesn't already start with the app's name.
Old way: "Untitled.txt - Smart Text Editor" => "Smart Text Editor - Untitled.txt - Smart Text Editor"
New way: "Smart Text Editor - Untitled.txt" => "Smart Text Editor - Untitled.txt"
```

Part of the adjusted Window Controls Overlay handling, the `devicePixelRatio` style handling is now local to the app's omnibox, rather than the whole page. I didn't figure out the issue of it being scaled incorrectly on Retina displays, so that's gonna be for a different work session. I'll just stay with the localized code for now at least.

I also removed the `STE.appearance.titleBarAreaInsets` getter, as it looks like I wasn't actually using it anywhere in the codebase. I think I won't need it for anything down the line either, since I can use the API-based ones that come with WCO anyways. Now I think I better understand how they manage the title bar, thinking of it as a box of usable space, rather than insets that cover up the page. I liked that model originally, since I had just started to implement the Safe Area Insets styling for iOS (mobile), with `viewport-fit=cover`. Now that things are more available in CSS-land, it's not as bad to manage things directly in the CSS instead.