Editor Close Animations
Found the missing implementations for the Editor tab closing animations, which broke with the reworks from the last few updates. It's about 85% fixed, as the animation will now play for all closing states now (yay!), but it actually needs to *not* run the animation if a new Editor is replacing an auto-generated one. So I am trying to figure out why that's happening now.

The previous no-animation issue was caused from the animation attribute not being added to the `<body>`, when it should have been, and the transition duration was being calculated before it should have been, resulting in the duration being `0`.