Enigmatic Glue Code
Got everything working again! Now the Editor opening sequences work correct in all situations once again, and now syncing the app's state with it's Local Storage settings now works back and forth, once again.

Finding out that there are still hidden gems of wonk sprinkled throughout the codebase, even with the "type-safety". I think it was mostly due to my `any` typing for the `STE.settings` object though, that broke some things and caused some hidden unintended consequences for me.

So currently the values returned from `STE.settings.get()` are always either `string | undefined`, and sometimes I was checking them as bare values, like `boolean`, which works fine in non-strict mode, with `==` and `!=` checks (I didn't know I was specifically relying on this for getting things to work). So, it was a big collection of things. I moved the script to run in ESM (enabling strict-mode), the `STE.settings` object wasn't type-safe enough, hiding the type checking error, and the fact that the `STE.settings` object always returns strings is not something I like XD

Ok, and for the Editor opening issues, it looks like it was because the Editor constructor code is specifically relying on how certain things are called, which I hadn't noticed when refactoring them. So, I reorganized things a bit to make it read nicer, but it turns out that it only runs correctly in the original order, because some calls rely on other things to have already happened. Now that it's working again at least, I know what that order is, and I can abstract things out so it's easier to see what each stage of the constructor is, at least.

Part of that, I can now make the `Editor` class into an extension of the `NumTextElement` component itself, allowing me to hopefully be able to remove the `STE.query()` function, as I can instead just access the Editor's components directly from the element's properties itself, instead of querying all over the document using it's identifier.

I love how much you really don't know when you first start out, some of these old design decisions are completely baffling to me how, and it's hilarous XD

Either way, I'm still very proud of how much I was able to pull out of the original codebase, with my limited experience so far too. STE is and has been very capable with lots of great features and ideas, I just hadn't learned about the best design decisions to build it yet. I'm also very happy with having things like Git, because I can see how this donkey show has progressed over the years XD (tried to think of a weird phrase to explain how goofy STE is; oops, it's a bad thing that already existings! my bad XD) It's very cool to be able to see how my opinions on building things has changed over time, and see which things have and haven't worked for me, and what I ended up going with for a given roadblock. I think messages like these are very awesome to have too, it's really neat to be able to reflect on things so close to what you're actually working on. It's tightly coupled with the actual project and you can see how things correlate together.

Listening to '40 years', 'A Mouth of God' by Marco Minnemann, and I listened to 'The Scambot Holiday Special' - Mike Keneally just before this; both awesome experimental prog albums :)

Oh yeah, and Lunatic Soul 'Through Shaded Woods' before those two :)

Gonna try to push the current state of this branch, to the main branch. Everything *should* be ready for that now, but you never know 'till you try! Currently the main branch and live site are running the 'Workspace Module' commit.