NBTify CDN
Remembered that I'm still temporarily loading NBTify directly from my `node_modules/` folder, so I added a CDN fallback to JSDelivr for it.

This is also a temporary fallback, as I am going to redo the file reading and writing system to go through a single function, as currently it's spread throughout the app in multiple spots, meaning I'd have to implement NBT opening in more than one section of the codebase, which I'm not gonna do. I can make it better now! :D