Load Event Fix
Not totally sure what's causing it not to fire now, but the `window.addEventListener("load");` call isn't being dispatched once the page loads. This completely prevented the app from loading as a PWA since this is where the Web App Manifest and Service Worker are managed to be run.

I'm curious if it's not working because it's inside a `<script type="module">`, and maybe that is running after the `"load"` event has already fired? I don't think this can be the case though, so I'm not sure about that.

Since I can depend on the script running after the document has loaded though, I decidede to try moving the function handler's content into an async-IIFE, so I can keep the early-`return` call setup.

I tried to debug this theory in STE, only to find out that the `"load"` event doesn't fire in `<iframe>` contexts? Didn't know that! The load event did fire from within a `module` script inside of a regular tab context though, so it may be something else. Still haven't quite figured it out, but I do feel like it's something related to the module loading process maybe.