TypeScript && GitHub Pages!
Alright, we finally made it! I found out how to deploy a TypeScript app to GitHub Pages using a GitHub Action. Now that that is possible for me to do, I think it's the next step that STE probably needs to be able to move forward. So happy to have it working!

It's also now a much better time than when I tried to do it before, because I already made STE fully type-safe/typed-out, so it's not like I'm moving to TypeScript, *and* trying to type out the whole codebase all at once (that's what I tried to do the first time). JSDoc+JS is a great way to transition your app to TypeScript, it sets you up so much more so you have a clear cut path for migration :)