TypeScript + Module Import Maps
A great next step! Now everything is moved over to TypeScript, and I also added support for the Import Maps spec, using the ES Module Shims polyfill for browsers that don't support it quite yet. It's too cool, and the shim works too well to not try it out. It's great for having Node-like module resolution, which is very helpful for TypeScript also, since it doesn't currently provide a way to change the module identifier to a CDN URL when compiled to JS. I wish that was an option, I'd love to be able to do that. I'd probably take that over using ES Module Shims, as I just want a way to use TypeScript/ESM JS without a bundler.