Num-Text npm
This is preparing for moving my classic version of Num Text to npm, so I don't have to depend on loading it from a GitHub Pages deployment anymore, haha. That setup has been holding me in a bunch while trying to move STE and my other projects to be more module-based, and to rely on packages more explicitly, and semantically versioned. This is an important step in regards to moving STE to a build step, as relying on GitHub Pages to host one of my modules, without versioning, is very dangerous in terms of having things be dependable, and not falling apart on feature changes. This stack decision ultimately cost me of having a clean Git versioning history for STE, as there will no longer be a way to go back to old versions without re-gathering specific versions of my modules manually, which really sucks. I didn't know any better back then, or at least I didn't think that would really be an issue. Lol, now I know for sure :)