Service Workin'
Brought over the modern async logic from my Service Worker over in Dovetail. It was much nicer, and I really like the way of building it's functionality with functions, as for some reason that wasn't how I programmed a while back, haha. Not that I would know to build it like that to start with, it's just funny when you look back and things seem so obvious once you have more experience since when you first started out.

Made some other nice changes to the rest of the codebase too. Found out in VSCode that using dashed with JSDoc descriptions isn't actually fully supported to render only a single dash, so it's showing two in some places. I didn't like that, so I'm going to move my code that does that over to not having the dashes, instead. I think some of my other projects may do that as well, so I will look out for it as I go. I like the code readability with the dashes, but I'd rather it not instead break how the actual descriptions look in your editor, that's more important than looking at them from the code side. Ideally both would work, but yeah.