Vite Build!!
Bringing even more over from Dovetail! Got Vite working over there last week I think it was, and it really was a good step towards understanding how to set things up to work properly with a bundler. This is very nice for doing more with STE, as now I can start using npm packages directly, without any module resolution issues! Now it will bundle them as part of the app's code. Of course, I probably won't change my way of doing things just because of that, but it certainly opens the doors of way more things that are viably possible and actionable now. Did a little demo project with React, TypeScript, and Vite the other day, and it was actually fairly simple to get working! I really liked it. The only thing I didn't like was the bundle fluff that adding React added to the app size. Not that it's that big of a deal, since it's easy to add it now, but made me understand how people say it can be a bit overkill to add sometimes when your project can simply work great with the vanilla web stack. if you need more though, than go ahead. I think the noteworthy part is that sometimes people will choose the steamroller to put the tent stake in the ground. Maybe that's a bit overkill of a comparison, but I guess that follows it's own point, so I like it hehe.

Here's an article I read about that, the emphasis part being "Would you use an industrial-strength data engineering framework to schlep around a couple of 4MB spreadsheets?".

https://rhetoricize.medium.com/when-you-dont-need-react-934e20baf636

Listened to a lot of music today, also revisiting "Instrumental Math Rock Mix" tonight, I think I both either really like this playlist, or I'm in the coding zone finally (today's been challenging for that for some reason, I think it's because I got a brain road block on NBTify from working on it so much recently).

https://www.youtube.com/watch?v=3egXqihisFQ

My next step for STE and Dovetail is to get Vite config tooling working so I can have full TypeScript Service Workers working! That will be sick 🔥